Refactor backend to use open OSINT services
