+++ README.md
@@ -6,7 +6,7 @@
 
 - **تحليل رقم الهاتف** باستخدام مكتبة `phonenumbers`.
 - **ربط الحسابات الاجتماعية** عبر أداة [Maigret](https://github.com/soxoj/maigret).
-- **فحص التسريبات الأمنية** بواسطة واجهة [HaveIBeenPwned](https://haveibeenpwned.com/).
+- **فحص التسريبات الأمنية** عبر مصادر مفتوحة مثل [scylla.sh](https://scylla.sh) و[dehashed](https://www.dehashed.com) (قراءة فقط).
 - **تسجيل استعلامات المستخدم** في مجلد `logs/`.
 - **واجهة تفاعلية** مبنية بـReact وTailwindCSS.
 - **تصدير التقارير** بصيغة JSON أو PDF.
@@ -58,10 +58,6 @@ npm run dev
 ## 📁 ملفات التكوين والمفاتيح
 
 ```bash
-# backend/.env (موجود بالفعل، عدله بالقيم الصحيحة)
-HIBP_API_KEY=your_hibp_key
-NUMVERIFY_API_KEY=your_numverify_key
-
 # frontend/.env
 VITE_API_BASE=http://localhost:8000/api
 ```
