+++ backend/.env
@@ -1,2 +1,1 @@
-HIBP_API_KEY=your_hibp_key
-NUMVERIFY_API_KEY=your_numverify_key
+# Example environment file - no API keys needed
