+++ backend/.env.example
@@ -1,2 +1,1 @@
-HIBP_API_KEY=your_hibp_key
-NUMVERIFY_API_KEY=your_numverify_key
+# No external API keys are required
